% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lga_attributes documentation.R
\docType{data}
\name{lga_attributes}
\alias{lga_attributes}
\title{NSW Local Government Area classifications}
\format{
\subsection{\code{lga_attributes}}{

A data frame with 4 columns:
\describe{
\item{statistical_area}{LGA statistical area code}
\item{LGA_name}{Text names for NSW LGAs}
\item{OLG_group}{Office of Local Government group classification}
\item{council_type}{Council type classification}
...
}
}
}
\source{
\url{https://www.yourcouncil.nsw.gov.au/wp-content/uploads/2020/05/Australian-Classification-of-Local-Government-and-OLG-group-numbers.pdf}
}
\usage{
lga_attributes
}
\description{
Classification of NSW Local Government Areas by the Office of Local Government (OLG).
Sourced from the Australian Classification of Local Government and OLG group numbers.

Manually converted to Excel by Emma Whitnall on 1 Oct 2024.
}
\keyword{datasets}
